+++ src/main/java/com/gaussic/model/dcs_history/H190Pojo.java
@@ -7,20 +7,21 @@ import java.util.Objects;
 @Entity
 @Table(name = "h_190", schema = "wind", catalog = "")
 public class H190Pojo {
-    private int id;
+    private Long id;
     private Timestamp vTime;
     private Float v;
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
+
 
     @Basic
     @Column(name = "v_time")
